Separate queryable object so we can build multiple queries at the same time
